Buffer overrun when referencing 'curtime' field (rev. 12_b5)
Former-commit-id: 5a7db80e53a956750b95bc0f409f00a95d28b315

+++ src/glutil.h
@@ -99,7 +99,7 @@ char b_glob[MAX_EXEC_STR + 1];
 uint64_t glob_ui64_stor[MAX_GLOB_STOR_AR_COUNT];
 int64_t glob_si64_stor[MAX_GLOB_STOR_AR_COUNT];
 float glob_float_stor[MAX_GLOB_STOR_AR_COUNT];
-int32_t glob_curtime;
+int64_t glob_curtime;
 
 char *GLOBAL_PREEXEC;
 char *GLOBAL_POSTEXEC;
